Crates
Added crates - a script for the crate object controlling the sprite and a script for the player (GrabController) in order to hold crates. Notably, the crates no longer have the colored object class - this is a workaround due to the fact that I could not understand layer masks to implement this properly.